# perf: Various performance tuning