From #288, I cherry-picked only the ones that were not controversial and did not change the test significantly.

### When should the Request object be extended?

Since there is no Request object in node, we are currently calling extendRequestPrototype when we create an instance of Hono (assuming it will have been added to globalThis by then if we are testing).

If we "tweak benchmarks/handle-event/index.js", etc., and decide that globalThis.Request should be initialized before "hono.ts" is imported, then it could be done on import.

### What I did not cherry-pick.

These will be discussed in another PR.

* 2e2d443
* ca92ca9f69a60028ac22316

```
minimal handler x 348,756 ops/sec ±6.81% (70 runs sampled)
hono - trie-router(default) x 155,222 ops/sec ±3.98% (82 runs sampled)
hono - regexp-router x 182,453 ops/sec ±5.75% (79 runs sampled)
itty-router x 85,700 ops/sec ±3.59% (86 runs sampled)
sunder x 122,709 ops/sec ±1.57% (90 runs sampled)
worktop x 64,828 ops/sec ±5.27% (78 runs sampled)
Fastest is minimal handler
```