@yusukebe 2022-06-02
Hi @usualoma 

Thank you! It's good that's is a minimal change and almost tests have not been changed.

> If we "tweak benchmarks/handle-event/index.js", etc., and decide that globalThis.Request should be initialized before "hono.ts" is imported, then it could be done on import.

In current PR, `extendRequestPrototype` is called in the constructor on `Hono`, so this code is OK?

```ts
import { Hono } from 'hono'

globalThis.Request = Request
globalThis.Response = Response

const app = new Hono() // <--- extends `Request`
```

There's not a problem, I will merge it!